hotshot_new_protocol::coordinator::coordinator_builder

Trait State

pub trait State: Sized {

    type MembershipCoordinator;
    type Consensus;
    type Network;
    type StateManager;
    type Client;
    type VidDisperser;
    type VidReconstructor;
    type VidFragmentAccumulator;
    type Vote1Collector;
    type Vote2Collector;
    type TimeoutCollector;
    type TimeoutOneHonestCollector;
    type EpochRootCollector;
    type CertVerifiers;
    type EpochManager;
    type BlockBuilder;
    type ProposalValidator;
    type ShareValidator;
    type Storage;
    type Outbox;
    type CoordinatorOutbox;
    type PublicKey;
    type NodeId;
    type Timer;
    type Fetcher;
    type Server;
    type Metrics;
    type Participation;

}

Builder’s type state specifies if members are set or not (unset).

You can use the associated types of this trait to control the state of individual members with the IsSet and IsUnset traits. You can change the state of the members with the Set* structs available in this module.

Required Associated Types

type MembershipCoordinator

Type state of the member membership_coordinator.

It can implement either IsSet or IsUnset

type Consensus

Type state of the member consensus.

It can implement either IsSet or IsUnset

type Network

Type state of the member network.

It can implement either IsSet or IsUnset

type StateManager

Type state of the member state_manager.

It can implement either IsSet or IsUnset

type Client

Type state of the member client.

It can implement either IsSet or IsUnset

type VidDisperser

Type state of the member vid_disperser.

It can implement either IsSet or IsUnset

type VidReconstructor

Type state of the member vid_reconstructor.

It can implement either IsSet or IsUnset

type VidFragmentAccumulator

Type state of the member vid_fragment_accumulator.

It can implement either IsSet or IsUnset

type Vote1Collector

Type state of the member vote1_collector.

It can implement either IsSet or IsUnset

type Vote2Collector

Type state of the member vote2_collector.

It can implement either IsSet or IsUnset

type TimeoutCollector

Type state of the member timeout_collector.

It can implement either IsSet or IsUnset

type TimeoutOneHonestCollector

Type state of the member timeout_one_honest_collector.

It can implement either IsSet or IsUnset

type EpochRootCollector

Type state of the member epoch_root_collector.

It can implement either IsSet or IsUnset

type CertVerifiers

Type state of the member cert_verifiers.

It can implement either IsSet or IsUnset

type EpochManager

Type state of the member epoch_manager.

It can implement either IsSet or IsUnset

type BlockBuilder

Type state of the member block_builder.

It can implement either IsSet or IsUnset

type ProposalValidator

Type state of the member proposal_validator.

It can implement either IsSet or IsUnset

type ShareValidator

Type state of the member share_validator.

It can implement either IsSet or IsUnset

type Storage

Type state of the member storage.

It can implement either IsSet or IsUnset

type Outbox

Type state of the member outbox.

It can implement either IsSet or IsUnset

type CoordinatorOutbox

Type state of the member coordinator_outbox.

It can implement either IsSet or IsUnset

type PublicKey

Type state of the member public_key.

It can implement either IsSet or IsUnset

type NodeId

Type state of the member node_id.

It can implement either IsSet or IsUnset

type Timer

Type state of the member timer.

It can implement either IsSet or IsUnset

type Fetcher

Type state of the member fetcher.

It can implement either IsSet or IsUnset

type Server

Type state of the member server.

It can implement either IsSet or IsUnset

type Metrics

Type state of the member metrics.

It can implement either IsSet or IsUnset

type Participation

Type state of the member participation.

It can implement either IsSet or IsUnset

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.