hotshot_new_protocol::coordinator

Module coordinator_maker

Tools for manipulating the type state of CoordinatorMaker.

See the detailed guide that describes how all the pieces here fit together.

Modules

sealed 🔒

Structs

Empty

Represents a State that has IsUnset implemented for all members.

SetConsensusMetrics

Represents a State that has IsSet implemented for State::ConsensusMetrics.

SetInitializer

Represents a State that has IsSet implemented for State::Initializer.

SetLockedQc

Represents a State that has IsSet implemented for State::LockedQc.

SetMembershipCoordinator

Represents a State that has IsSet implemented for State::MembershipCoordinator.

SetMetrics

Represents a State that has IsSet implemented for State::Metrics.

SetNetwork

Represents a State that has IsSet implemented for State::Network.

SetPrivateKey

Represents a State that has IsSet implemented for State::PrivateKey.

SetPublicKey

Represents a State that has IsSet implemented for State::PublicKey.

SetStakeTableCapacity

Represents a State that has IsSet implemented for State::StakeTableCapacity.

SetStatePrivateKey

Represents a State that has IsSet implemented for State::StatePrivateKey.

SetStorage

Represents a State that has IsSet implemented for State::Storage.

SetTimeoutDuration

Represents a State that has IsSet implemented for State::TimeoutDuration.

SetUpgradeLock

Represents a State that has IsSet implemented for State::UpgradeLock.

Traits

IsComplete

Marker trait that indicates that all required members are set.

IsSet

Marker trait that indicates that the member is set, i.e. at least one of its setters was called.

IsUnset

Marker trait that indicates that the member is unset, i.e. none of its setters was called.

State

Builder’s type state specifies if members are set or not (unset).