hotshot_new_protocol::coordinator::coordinator_maker

Trait State

pub trait State: Sized {

    type MembershipCoordinator;
    type Network;
    type Initializer;
    type UpgradeLock;
    type PublicKey;
    type PrivateKey;
    type StatePrivateKey;
    type StakeTableCapacity;
    type TimeoutDuration;
    type Storage;
    type Metrics;
    type ConsensusMetrics;
    type LockedQc;

}

Builder’s type state specifies if members are set or not (unset).

You can use the associated types of this trait to control the state of individual members with the IsSet and IsUnset traits. You can change the state of the members with the Set* structs available in this module.

Required Associated Types

type MembershipCoordinator

Type state of the member membership_coordinator.

It can implement either IsSet or IsUnset

type Network

Type state of the member network.

It can implement either IsSet or IsUnset

type Initializer

Type state of the member initializer.

It can implement either IsSet or IsUnset

type UpgradeLock

Type state of the member upgrade_lock.

It can implement either IsSet or IsUnset

type PublicKey

Type state of the member public_key.

It can implement either IsSet or IsUnset

type PrivateKey

Type state of the member private_key.

It can implement either IsSet or IsUnset

type StatePrivateKey

Type state of the member state_private_key.

It can implement either IsSet or IsUnset

type StakeTableCapacity

Type state of the member stake_table_capacity.

It can implement either IsSet or IsUnset

type TimeoutDuration

Type state of the member timeout_duration.

It can implement either IsSet or IsUnset

type Storage

Type state of the member storage.

It can implement either IsSet or IsUnset

type Metrics

Type state of the member metrics.

It can implement either IsSet or IsUnset

type ConsensusMetrics

Type state of the member consensus_metrics.

It can implement either IsSet or IsUnset

type LockedQc

Type state of the member locked_qc.

It can implement either IsSet or IsUnset

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.