hotshot_new_protocol::coordinator

Module coordinator_builder

Tools for manipulating the type state of CoordinatorBuilder.

See the detailed guide that describes how all the pieces here fit together.

Modules

sealed 🔒

Structs

Empty

Represents a State that has IsUnset implemented for all members.

SetBlockBuilder

Represents a State that has IsSet implemented for State::BlockBuilder.

SetCheckpointCollector

Represents a State that has IsSet implemented for State::CheckpointCollector.

SetConsensus

Represents a State that has IsSet implemented for State::Consensus.

SetEpochManager

Represents a State that has IsSet implemented for State::EpochManager.

SetMembershipCoordinator

Represents a State that has IsSet implemented for State::MembershipCoordinator.

SetNetwork

Represents a State that has IsSet implemented for State::Network.

SetNodeId

Represents a State that has IsSet implemented for State::NodeId.

SetOutbox

Represents a State that has IsSet implemented for State::Outbox.

SetProposalValidator

Represents a State that has IsSet implemented for State::ProposalValidator.

SetPublicKey

Represents a State that has IsSet implemented for State::PublicKey.

SetStateManager

Represents a State that has IsSet implemented for State::StateManager.

SetTimeoutCollector

Represents a State that has IsSet implemented for State::TimeoutCollector.

SetTimeoutOneHonestCollector

Represents a State that has IsSet implemented for State::TimeoutOneHonestCollector.

SetTimer

Represents a State that has IsSet implemented for State::Timer.

SetVidDisperser

Represents a State that has IsSet implemented for State::VidDisperser.

SetVidReconstructor

Represents a State that has IsSet implemented for State::VidReconstructor.

SetVote1Collector

Represents a State that has IsSet implemented for State::Vote1Collector.

SetVote2Collector

Represents a State that has IsSet implemented for State::Vote2Collector.

Traits

IsComplete

Marker trait that indicates that all required members are set.

IsSet

Marker trait that indicates that the member is set, i.e. at least one of its setters was called.

IsUnset

Marker trait that indicates that the member is unset, i.e. none of its setters was called.

State

Builder’s type state specifies if members are set or not (unset).